extract http request data reading logic
