+++ src/webserver/mod.rs
@@ -1,6 +1,7 @@
 pub mod database;
 pub mod error_with_status;
 pub mod http;
+pub mod http_request_info;
 
 pub use database::Database;
 pub use error_with_status::ErrorWithStatus;
